requirements/base.txt
pytz==2021.3  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==9.0.1  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.1.3  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.2.3  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.2.1  # https://github.com/celery/django-celery-beat
flower==1.0.0  # https://github.com/mher/flower
django-filter
# Django
# ------------------------------------------------------------------------------
django==3.2.12  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.48.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.13.1  # https://github.com/encode/django-rest-framework
django-cors-headers==3.11.0 # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.21.2

# Swagger
drf-nested-routers==0.93.4
drf-yasg==1.20.0

# JWT
djangorestframework-simplejwt==5.0.0

# Django Rest Auth
django-rest-auth==0.9.5